requirements.txt
anyio==4.9.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
notion-client==2.3.0
python-dotenv==1.1.0
sniffio==1.3.1
typing_extensions==4.13.1
